fix: pr improvements
