+++ packages/sdk-shared/coverage-tests/custom/TestCheckDomSnapshotCrossOriginFrames_VG.spec.js
@@ -10,15 +10,14 @@ describe('Coverage Tests', () => {
   let driver, eyes, serverA, serverB, url
 
   beforeEach(async () => {
-    url = adjustUrlToDocker('http://localhost:7373/handles/cors_frames/cors.hbs')
+    url = adjustUrlToDocker('http://localhost:7373/cors_frames/cors.hbs')
     const staticPath = path.join(__dirname, '../fixtures')
     serverA = await testServer({
       port: 7373,
       staticPath,
       allowCors: false,
-      middlewareFile: path.resolve(__dirname, '../util/cors-snapshot-middleware.js'),
+      middlewareFile: path.resolve(__dirname, '../util/handlebars-middleware.js'),
       hbData: {
-        title: 'Hello, I have a designer cross origin frame',
         src: adjustUrlToDocker('http://localhost:7374/cors_frames/frame.html'),
       },
     })
