+++ packages/sdk-shared/coverage-tests/fixtures/cors_frames/cors.hbs
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <body>
-    <h1>{{ title }}</h1>
+    <h1>Hello, I have a designer cross origin frame</h1>
     <iframe src="{{ src }}"
         style="display: flex; width: 100%; height: 100%; flex-direction: column;"></iframe>
 </body>
