--- packages/sdk-shared/coverage-tests/util/cors-snapshot-middleware.js
@@ -1,18 +1,0 @@
-const Handlebars = require('handlebars')
-const {readFileSync} = require('fs')
-
-const getFilePath = (url, staticPath) => {
-  return `${staticPath}${url.replace('/handles', '')}`
-}
-
-module.exports = ({hbData, staticPath}) => (req, res, next) => {
-  if (/\.hbs$/.test(req.url)) {
-    const data = JSON.parse(hbData)
-    const filePath = getFilePath(req.url, staticPath)
-    const file = readFileSync(filePath).toString()
-    const compiled = Handlebars.compile(file)(data)
-    res.send(compiled)
-  } else {
-    next()
-  }
-}
